# spec/openapi-spec.yml
openapi: 3.0.3
info:
  title: In The Mealtime
  description: This is the API for In The Mealtime
  version: 1.0.0
...
paths:
  /api/order:
    get:
      tags:
        - Order
      summary: Fetch all orders
      description: Fetch all orders
      operationId: fetchOrders
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
  /api/order/{id}:
    get:
      tags:
        - Order
      summary: Fetch single order
      description: Fetch single order
      operationId: fetchOrder
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      responses:
        200:
          $ref: "#/components/responses/SuccessOrderResponse"
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
    delete:
      tags:
        - Order
      summary: Deletes as single order
      description: Deletes as single order
      operationId: deleteOrder
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ID"
        - in: header
          name: If-Match
          required: true
          schema:
            $ref: "#/components/schemas/Version"
      responses:
        '204':
          description: Successful operation
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        412:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
  /api/order/restaurant:
    get:
      tags:
        - Order
      summary: Fetch orderable restaurants
      description: Fetch orderable restaurants
      operationId: fetchOrderableRestaurants
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ID'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ID'
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
  /api/order/restaurant/{restaurantId}:
    post:
      tags:
        - Order
      summary: Create new order for a certain restaurant
      description: Create new order for a certain restaurant
      operationId: createOrder
      parameters:
        - in: path
          name: restaurantId
          required: true
          schema:
            $ref: "#/components/schemas/ID"
          description: ID of the restaurant for that the order will be
      responses:
        200:
          $ref: "#/components/responses/SuccessOrderResponse"
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
  /api/order/{orderId}/infos:
    put:
      tags:
        - Order
      summary: Sets Infos to an Order
      description: Sets Infos to an Order
      operationId: setOrderInfo
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            $ref: "#/components/schemas/ID"
        - in: header
          name: If-Match
          required: true
          schema:
            $ref: "#/components/schemas/Version"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInfosPatch"
          application/xml:
            schema:
              $ref: "#/components/schemas/OrderInfosPatch"
      responses:
        200:
          $ref: "#/components/responses/SuccessOrderResponse"
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        412:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
  /api/order/{orderId}/order-position:
    post:
      tags:
        - Order
      summary: Creates a new order position
      description: Creates a new order position
      operationId: createOrderPosition
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderPositionPatch"
          application/xml:
            schema:
              $ref: "#/components/schemas/OrderPositionPatch"
      responses:
        200:
          $ref: "#/components/responses/SuccessOrderResponse"
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
  /api/order/{orderId}/order-position/{orderPositionId}:
    patch:
      tags:
        - Order
      summary: Updates an order position
      description: Updates an order position
      operationId: updateOrderPosition
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            $ref: "#/components/schemas/ID"
        - in: path
          name: orderPositionId
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderPositionPatch"
          application/xml:
            schema:
              $ref: "#/components/schemas/OrderPositionPatch"
      responses:
        200:
          $ref: "#/components/responses/SuccessOrderResponse"
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
    delete:
      tags:
        - Order
      summary: Deletes an order position
      description: Deletes an order position
      operationId: deleteOrderPosition
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            $ref: "#/components/schemas/ID"
        - in: path
          name: orderPositionId
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
            application/xml:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
  /api/order/{id}/lock:
    patch:
      tags:
        - Order
      summary: Locks Order to try ordering
      description: Locks Order to try ordering
      operationId: lockOrder
      parameters:
        - in: path
          required: true
          name: id
          schema:
            $ref: "#/components/schemas/ID"
        - in: header
          name: If-Match
          required: true
          schema:
            $ref: "#/components/schemas/Version"
      responses:
        200:
          $ref: "#/components/responses/SuccessOrderResponse"
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        412:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
    delete:
      tags:
        - Order
      summary: Re-open Order after locking
      description: Re-open Order after locking
      operationId: reopenOrder
      parameters:
        - in: path
          required: true
          name: id
          schema:
            $ref: "#/components/schemas/ID"
        - in: header
          name: If-Match
          required: true
          schema:
            $ref: "#/components/schemas/Version"
      responses:
        200:
          $ref: "#/components/responses/SuccessOrderResponse"
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        412:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
  /api/order/{id}/ordered:
    patch:
      tags:
        - Order
      summary: Set Order to be ordered and waiting for delivery
      description: Set Order to be ordered and waiting for delivery
      operationId: orderIsNowOrdered
      parameters:
        - in: path
          required: true
          name: id
          schema:
            $ref: "#/components/schemas/ID"
        - in: header
          name: If-Match
          required: true
          schema:
            $ref: "#/components/schemas/Version"
      responses:
        200:
          $ref: "#/components/responses/SuccessOrderResponse"
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        412:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
...
  /api/order/{id}/revoke:
    delete:
      tags:
        - Order
      summary: Revoke Order
      description: Revoke Order
      operationId: revokeOrder
      parameters:
        - in: path
          required: true
          name: id
          schema:
            $ref: "#/components/schemas/ID"
        - in: header
          name: If-Match
          required: true
          schema:
            $ref: "#/components/schemas/Version"
      responses:
        200:
          $ref: "#/components/responses/SuccessOrderResponse"
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        412:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
  /api/order/{id}/archive:
    delete:
      tags:
        - Order
      summary: Archive Order
      description: Archive Order
      operationId: archiveOrder
      parameters:
        - in: path
          required: true
          name: id
          schema:
            $ref: "#/components/schemas/ID"
        - in: header
          name: If-Match
          required: true
          schema:
            $ref: "#/components/schemas/Version"
      responses:
        200:
          $ref: "#/components/responses/SuccessOrderResponse"
        400:
          $ref: "#/components/responses/ClientErrorResponse"
        412:
          $ref: "#/components/responses/ClientErrorResponse"
        500:
          $ref: "#/components/responses/ServerErrorResponse"
...
components:
  schemas:
...
    Order:
      type: object
      required:
        - id
        - restaurantId
        - infos
        - orderPositions
        - orderState
        - stateManagement
      allOf:
        - $ref: "#/components/schemas/Updatable"
      properties:
        id:
          $ref: "#/components/schemas/ID"
        restaurantId:
          $ref: "#/components/schemas/ID"
        infos:
          $ref: "#/components/schemas/OrderInfos"
        orderState:
          $ref: "#/components/schemas/OrderStateType"
        date:
          type: string
          format: date
        stateManagement:
          $ref: "#/components/schemas/OrderStateManagement"
        orderPositions:
          type: array
          items:
            $ref: "#/components/schemas/OrderPosition"
    OrderStateManagement:
      type: object
      properties:
        locked_at:
          type: string
          format: date-time
        ordered_at:
          type: string
          format: date-time
        delivered_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
        archived_at:
          type: string
          format: date-time
        next_transition_duration:
          type: string
          format: duration
        next_transition_timestamp:
          type: string
          format: date-time
    OrderInfosPatch:
      type: object
      properties:
        orderer:
          type: string
        fetcher:
          type: string
        moneyCollectionType:
          $ref: "#/components/schemas/OrderMoneyCollectionType"
        moneyCollector:
          type: string
        orderClosingTime:
          type: string
          format: time
        orderText:
          type: string
        maximumMealCount:
          type: integer
          format: int32
          minimum: 1
    OrderInfos:
      allOf:
        - $ref: "#/components/schemas/OrderInfosPatch"
    OrderPositionPatch:
      required:
        - name
        - meal
        - price
      properties:
        name:
          type: string
        meal:
          type: string
        price:
          type: number
          format: float
        paid:
          type: number
          format: float
        tip:
          type: number
          format: float
    OrderPosition:
      type: object
      allOf:
        - $ref: "#/components/schemas/OrderPositionPatch"
      required:
        - id
        - name
        - meal
        - price
        - index
      properties:
        id:
          $ref: "#/components/schemas/ID"
        index:
          type: integer
          format: int32
    OrderStateType:
      type: string
      enum:
        - New
        - Open
        - Locked
        - Revoked
        - Ordered
        - Delivered
        - Archived
    OrderMoneyCollectionType:
      type: string
      enum:
        - Bar
        - PayPal
    ID:
      type: string
      format: uuid
    Version:
      type: string
      format: uuid
    ErrorObject:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        correlationId:
          type: string
        message:
          type: string
        detailMessage:
          type: string
        stacktrace:
          type: string
    Creatable:
      type: object
      required:
        - createdAt
        - createdBy
      properties:
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
    Updatable:
      type: object
      required:
        - version
      allOf:
        - $ref: "#/components/schemas/Creatable"
      properties:
        version:
          $ref: "#/components/schemas/Version"
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
  responses:
    SuccessOrderResponse:
      description: Successful operation
      headers:
        ETag:
          schema:
            $ref: "#/components/schemas/Version"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Order'
        application/xml:
          schema:
            $ref: '#/components/schemas/Order'
...
    ClientErrorResponse:
      description: Operation failed due to client error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorObject"
        application/xml:
          schema:
            $ref: "#/components/schemas/ErrorObject"
    ServerErrorResponse:
      description: Operation failed due to client error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorObject"
        application/xml:
          schema:
            $ref: "#/components/schemas/ErrorObject"